refactor: 전역 객체 은닉
